refactor(app): move all root directories to scopes
features/ -> @features/
lib/ -> @lib/
ui/ -> @ui/
