+++ README.md
@@ -1,0 +1,7 @@
+# @howtocards/frontend
+
+## Aliases
+
+- `@features` resolves to `./src/features/`
+- `@ui` resolves to `./src/ui/`
+- `@lib` resolves to `./src/lib/`
