+++ src/app.js
@@ -1,11 +1,11 @@
 import React from 'react'
 import { hot } from 'react-hot-loader'
 
-import { AccountLoader } from 'features/common'
+import { AccountLoader } from '@features/common'
 
-import { ToggleThemeProvider } from 'lib/theme-context'
-import { lightTheme } from 'ui/themes/light'
-import { darkTheme } from 'ui/themes/dark'
+import { ToggleThemeProvider } from '@lib/theme-context'
+import { lightTheme } from '@ui/themes/light'
+import { darkTheme } from '@ui/themes/dark'
 
 import { rootRoutes } from './routes'
