+++ src/features/cards/organisms/cards-list.js
@@ -2,7 +2,7 @@ import React from 'react'
 import PropTypes from 'prop-types'
 import styled from 'styled-components'
 
-import { ConditionalList } from 'ui/organisms'
+import { ConditionalList } from '@ui/organisms'
 
 
 const CardsItemsList = styled.div`
