+++ src/features/cards/pages/create.js
@@ -3,11 +3,11 @@ import PropTypes from 'prop-types'
 import { connect } from 'react-redux'
 import { compose } from 'recompose'
 import { withFormik } from 'formik'
-import { Col } from 'styled-components-layout'
 
-import { Authenticated } from 'features/common'
-import { Card, Input, Button } from 'ui/atoms'
-import { TextArea } from 'ui/molecules'
+import { Authenticated } from '@features/common'
+import { Card, Input, Button } from '@ui/atoms'
+import { TextArea } from '@ui/molecules'
+import { Col } from '@lib/styled-components-layout'
 import { CardsCommonTemplate } from '../templates/common'
 import { letterCreate } from '../effects'
