+++ src/features/common/pages/not-found.js
@@ -1,8 +1,8 @@
 import React from 'react'
 import styled from 'styled-components'
 
-import { CommonContentTemplate } from 'features/common'
-import { Container } from 'ui/templates'
+import { CommonContentTemplate } from '@features/common'
+import { Container } from '@ui/templates'
 
 
 const NotFoundContainer = styled.div`
