+++ src/features/join/pages/registration.js
@@ -4,10 +4,10 @@ import { connect } from 'react-redux'
 import { compose } from 'recompose'
 import { withFormik } from 'formik'
 
-import { Col, Row } from 'styled-components-layout'
-import { Card, Input, H2, ButtonPrimary, Link, ErrorBox } from 'ui/atoms'
-import { PrimitiveFooter } from 'ui/organisms'
-import { Container, CenterContentTemplate } from 'ui/templates'
+import { Col, Row } from '@lib/styled-components-layout'
+import { Card, Input, H2, ButtonPrimary, Link, ErrorBox } from '@ui/atoms'
+import { PrimitiveFooter } from '@ui/organisms'
+import { Container, CenterContentTemplate } from '@ui/templates'
 
 import { userRegister } from '../effects/registration'
 import { userLogin } from '../effects/join'
