+++ src/reducers.js
@@ -1,7 +1,7 @@
 import { combineReducers } from 'redux'
 
-import { commonReducer as common } from 'features/common'
-import { cardsReducer as cards } from 'features/cards'
+import { commonReducer as common } from '@features/common'
+import { cardsReducer as cards } from '@features/cards'
 
 
 export const rootReducer = combineReducers({
