+++ src/routes.js
@@ -1,9 +1,9 @@
 import React from 'react'
 import { renderRoutes } from 'react-router-config'
 
-import { NotFoundPage } from 'features/common'
-import { joinRoutes } from 'features/join'
-import { cardsRoutes } from 'features/cards'
+import { NotFoundPage } from '@features/common'
+import { joinRoutes } from '@features/join'
+import { cardsRoutes } from '@features/cards'
 
 
 const routes = [
