+++ src/ui/templates/container.js
@@ -1,5 +1,5 @@
 import styled from 'styled-components'
-import { mixins, WithTag } from 'styled-components-layout'
+import { mixins, WithTag } from '@lib/styled-components-layout'
 
 
 export const Container = styled(WithTag)`
